Some more lua nvim setup now that I'm actually getting my head around it
Packer requires PackerCompile and PackerInstall to be run manually and
then nvim turned off/on. Compile errors can really break things too.

If you use a config file, it gets loaded everytime, as opposed to
compiled. This is a good way to not have to use PackerCompile when
playing around with non-install based settings